Refactor package name for axios.js
