+++ sticky-notes-fullstack/sticky-notes-frontend/src/components/auth/Register.js
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 
-import axios from "./axios";
+import axios from "../api/axios";
 
 import styles from "./Register.module.css";
